first draft of github bug retriver.
GitHub limits the number of result per request, so in order to get all
issues we have to download several pages.

Everything is synchroneous right now for simplicity. I will tackle on
this problem later on.

This is a PoC. There is very little data imported from github but
we can see Nicolas' org-element-interpret-data already works :)