first draft of import/sync.
Run org-sync-import in a document that doesn't have a buglist headline
to insert a "Buglist" headline with all the associated bugs.

Alternatively, add the following in your document:

    * Buglist
    :PROPERTIES:
    :url:      https://api.github.com/repos/octocat/Hello-World/issues
    :END:

and run org-sync-pull. It will insert (overwrite for now) the content
of this headline with the same buglist you would get with
org-sync-import.

No error checking atm.