first draft of the org-sync lib.
Org-sync provides several overridable functions that implement actions
and so that each backend can implement its own action.

The overridable action are defined with the os-defun-overridable macro
instead of defun. The name of function must start with "os--" followed
by the action name.

To override it, define a regular function with the name
"os-BACKEND-myaction". Every time os--myaction is called it first
looks at the current value of os-backend which is the symbol of the
current backend. If os-BACKEND-myaction exists, it is
called. Otherwise, the regular "base" function is called.

The os-backend-alist variable maps url patterns to backend symbols.
The function os-get-backend takes a url as input and performs a
look-up on os-backend-alist.

There is also the os-with-backend macro which is basically some sugar
that dynamically binds os-backend and does error-checking. It lets you do
things like:

    (os-with-backend my-url-or-backend-symbol
      (os--overrided-action))