os: os-headline-to-buglist calls os-headline-to-bug on its descendants
os-headline-to-buglist's paremeter is the headline element which
contains all bugs directly as its child headlines. the first element
of its contents is the section part with property drawer and text, all
other children are the headlines corresponding to bugs.

i don't understand why it could have worked the way it was.