Get gitlab server name from URL
This changes so that the server FQDN of the GitLab server is parsed from
the URL to the gitlab repository. This way we don't have to specify it
multiple times.

Further, in the backend list we're changing to match any string
containing "gitlab" to be associated with the gitlab backend, this
includes gitlab.com but also private instances that happen to have
gitlab in the URL (which is true for the one I'm using). I'm not sure
this last piece is correct or in line with out org-sync typically
wants to do things. I see that there is an assumption that any URL
containing /projects/ belong to redmine, which seems a little aggressive
but anyway there is no similar part of the URL which we can match on for
GitLab. How do we want to handle private repos? Right now a user with a
private GitLab instance would need to append an entry to the
org-sync-backend-alist.