Fix parsing of PROPERTIES drawer
(car org-elements-contents h) will get the first part of h and normally when the
PROPERTIES drawer is the first thing after the heading title we can get the
first of the first section and it will be the properties drawer (first heading,
then content of heading then in that we have the drawer). However, if there is a
DEADLINE or SCHEDULE in between we will get that instead. By not going so deep
and instead passing in the entire content of the section into
org-sync-property-drawer-to-alist we let it recursively search through for the
node-property list (what org-element calls the PROPERTIES drawer).

We're much less strict about where the PROPERTIES drawer goes which I think is
generally a good thing. It's certainly a good thing that it allows us to have a
SCHEDULE / DEADLINE line just under the heading. I hope this doesn't break
something else though.

Fixes #36.