Send issue data in body
We used to send the issue data like title and description in the URI but
there's a limit to the length of data we can send. By instead sending
the data JSON encoded in the body we get around this!

I introduced a second function for POST/PUTing JSON data that sets the
correct Content-Type. This can probably be written in a more elegant
manner but this works for now :)

Fixes #41.